Better handle redirect from login
